samples: nrf_desktop: Split battery charging from battery measurement
Move these features in separate modules, so they can be compiled
independently into different hardware configurations.

Jira: DESK-540
