+++ applications/nrf_desktop/configuration/nrf52840_pca20041/app_ZDebug.conf
@@ -23,15 +23,16 @@ CONFIG_DESKTOP_LED_ENABLE=y
 CONFIG_DESKTOP_LED_COUNT=2
 CONFIG_DESKTOP_LED_COLOR_COUNT=3
 
-CONFIG_DESKTOP_BATTERY_DISCRETE=y
-CONFIG_DESKTOP_BATTERY_CHARGE_ENABLE_PIN=14
-CONFIG_DESKTOP_BATTERY_MONITOR_ENABLE_PIN=6
-CONFIG_DESKTOP_BATTERY_CSO_PIN=8
-CONFIG_DESKTOP_BATTERY_CSO_PULL_UP=y
-CONFIG_DESKTOP_BATTERY_CSO_FREQ=1
+CONFIG_DESKTOP_BATTERY_CHARGER=y
+CONFIG_DESKTOP_BATTERY_CHARGER_ENABLE_PIN=14
+CONFIG_DESKTOP_BATTERY_CHARGER_CSO_PIN=8
+CONFIG_DESKTOP_BATTERY_CHARGER_CSO_PULL_UP=y
+CONFIG_DESKTOP_BATTERY_CHARGER_CSO_FREQ=1
 
-CONFIG_DESKTOP_BATTERY_MIN_LEVEL=3100
-CONFIG_DESKTOP_BATTERY_MAX_LEVEL=4200
+CONFIG_DESKTOP_BATTERY_MEAS=y
+CONFIG_DESKTOP_BATTERY_MEAS_ENABLE_PIN=6
+CONFIG_DESKTOP_BATTERY_MEAS_MIN_LEVEL=3100
+CONFIG_DESKTOP_BATTERY_MEAS_MAX_LEVEL=4200
 CONFIG_DESKTOP_VOLTAGE_TO_SOC_DELTA=10
 
 CONFIG_DESKTOP_USB_ENABLE=y
